+++ packages/job-queue/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @workglow/job-queue
 
+## 0.3.23
+
+### Refactors
+
+- consolidate hardcoded limits into DEFAULT_LIMITS/SECURITY_LIMITS (#609)
+
 ## 0.3.22
 
 ### Bug Fixes