+++ providers/huggingface-inference/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.3.23
+
+### Chores
+
+- format / lint
+
 ## 0.3.22
 
 ## 0.3.21